More features and fix bugs
Fixed bugs pertaining id of a page, which was being used as if it were the page number of a certain book. I had to rewrite the DDL and inserts for the database as well in order to improve the naming of the procedures and its parameters. In the Web API, I created more endpoints for every stored procedure so that we can do CRUD on books AND pages.